Refactor of folder layout
