+++ app/src/main/java/com/example/libraryapp/db/floors/Floors.java
@@ -1,4 +1,4 @@
-package com.example.libraryapp.db;
+package com.example.libraryapp.db.floors;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
